move remaining code out of enterprise pkg
