+++ client/web/src/enterprise/own/graphqlQueries.ts
@@ -45,19 +45,3 @@ export const DELETE_INGESTED_CODEOWNERS_MUTATION = gql`
         }
     }
 `
-
-export const ASSIGN_OWNER = gql`
-    mutation AssignOwner($input: AssignOwnerOrTeamInput!) {
-        assignOwner(input: $input) {
-            alwaysNil
-        }
-    }
-`
-
-export const ASSIGN_TEAM = gql`
-    mutation AssignTeam($input: AssignOwnerOrTeamInput!) {
-        assignTeam(input: $input) {
-            alwaysNil
-        }
-    }
-`
